login redirect

+++ www/admin/teachers/delete.php
@@ -3,6 +3,13 @@ require '../../../vendor/autoload.php';
 include '../../../autoload.php';
 
 use controllers\TeacherController;
+use core\helpers\HTTP;
+
+if (!isset($_SESSION['admin'])) {
+    HTTP::redirect("/login");
+    exit();
+}
+
 
 $teacherController = new TeacherController();
 $teacherController->deleteTeacher($_GET['id']);
